requirements.txt
pandas
numpy
matplotlib
torch
scikit-image
xarray
tqdm
openpyxl
torch
torchvision
h5py
ipykernel
requests
bs4
cartopy
netcdf4
pyyaml
pytest
dask
gpm-api
black
isort
hydra-core
omegaconf
h5netcdf
pytorch-lightning
tensorboard
einops
xskillscore